discount creation for admin and user implemented, primitive fields changed to class types, migrations in blit entity and discount entity added, excel service changes  applied, discount view model and common blit view model migration effects applied, in admin event service changing event state to ENDED now change event dates and blit types to CLOSED state. primitive to class types side effects fixed, discount service test fixed. discount integration tests passed. double amount changed to long value in and migration applied

+++ src/main/java/com/blito/models/Discount.java
@@ -24,7 +24,7 @@ public class Discount {
 	
 	private Double percentage;
 	
-	private Double amount;
+	private Long amount;
 	
 	@ManyToOne
 	@JoinColumn(name="userId")
@@ -105,11 +105,11 @@ public class Discount {
 		this.percentage = percentage;
 	}
 
-	public Double getAmount() {
+	public Long getAmount() {
 		return amount;
 	}
 
-	public void setAmount(Double amount) {
+	public void setAmount(Long amount) {
 		this.amount = amount;
 	}
